Minor scrutinizer fixes (#1226)
* Minor scrutinizer fixes

* Remove spurious debug output

+++ src/PhpSpreadsheet/Calculation/TextData.php
@@ -266,7 +266,7 @@ class TextData
         if (!is_numeric($value) || !is_numeric($decimals)) {
             return Functions::NAN();
         }
-        $decimals = floor($decimals);
+        $decimals = (int) floor($decimals);
 
         $valueResult = round($value, $decimals);
         if ($decimals < 0) {
